+++ src/vivarium_inputs/core.py
@@ -219,14 +219,14 @@ def _get_additional_id_columns(data, entities):
         Covariate: 'covariate_id',
         Risk: 'risk_id',
         Etiology: 'etiology_id',
-        CoverageGap: 'coverage_gap_id',
+        CoverageGap: 'coverage_gap',
         HealthcareEntity: 'healthcare_entity',
         HealthTechnology: 'health_technology',
     }
     out = set()
     out.add(id_column_map[type(entities[0])])
-    if isinstance(entities[0], CoverageGap) and data['measure'].all() in ['relative_risk', 'population_attributable_fraction']:
-        out.add('rei_id')
+    if isinstance(entities[0], CoverageGap) and data['measure'].all() in ['relative_risk']:
+        out.add('risk_id')
     out |= set(data.columns) & set(id_column_map.values())
     return out
 
@@ -268,6 +268,7 @@ def _validate_data(data: pd.DataFrame, key_columns: Iterable[str]=None):
 ####################
 # Cause-like stuff #
 ####################
+
 
 def _get_death(entities, location_ids):
     measure_ids = _get_ids_for_measure(entities, 'death')
@@ -393,6 +394,7 @@ def _get_disability_weight(entities, _):
         data.append(df)
 
     data = pd.concat(data, ignore_index=True)
+
     return data.reset_index(drop=True)
 
 
@@ -483,6 +485,7 @@ def _get_relative_risk(entities: Iterable[Union[Risk, CoverageGap]], location_id
             draw_cols = [f'draw_{i}' for i in range(1000)]
             measure_data.loc[:, draw_cols] = 1 / measure_data.loc[:, draw_cols]
             measure_data = _handle_special_coverage_gap_data(special_cases, measure_data, 1)
+            measure_data = measure_data.rename(columns={"rei_id": "risk_id"})
             df.append(measure_data)
 
         # any coverage_gap from aux_data
@@ -497,7 +500,8 @@ def _get_relative_risk(entities: Iterable[Union[Risk, CoverageGap]], location_id
                     data['year_id'] = id
                     missing_data = missing_data.append(data)
                 measure_data = pd.concat([measure_data] + missing_data)
-                measure_data['coverage_gap_id'] = np.NaN
+                measure_data['coverage_gap'] = entity.name
+                measure_data = measure_data.rename(columns={"rei_id": "risk_id"})
                 del measure_data['measure']
                 df.append(measure_data)
         measure_data = pd.concat(df)
@@ -537,7 +541,7 @@ def _compute_paf_for_special_cases(affected_entity: Union[Cause, Risk], entity: 
             rr = rr[(rr.cause_id == cause_id)]
         elif isinstance(affected_entity, Risk):
             risk_id, cause_id = affected_entity.gbd_id, np.nan
-            rr = rr[(rr.rei_id == risk_id)]
+            rr = rr[(rr.risk_id == risk_id)]
         else:
             raise InvalidQueryError(f'You requested the non-valid PAF data for {entity}-{affected_entity} pair')
 
@@ -554,13 +558,21 @@ def _compute_paf_for_special_cases(affected_entity: Union[Cause, Risk], entity: 
         temp = relative_risk[draw_columns]*exposure[draw_columns]
         temp_sum = temp.groupby(['age_group_id', 'year_id', 'sex_id']).sum()
         temp_result = ((temp_sum-1)/temp_sum)
+        temp_result = temp_result.replace(-np.inf, 0)  # Rows with zero exposure.
         temp_result['cause_id'] = cause_id
         temp_result['location_id'] = location_id
         temp_result['risk_id'] = risk_id
         temp_result['measure_id'] = 3
+        if isinstance(entity, CoverageGap):
+            temp_result['coverage_gap'] = entity.name
         paf.append(temp_result.reset_index())
 
     paf_data = pd.concat(paf)
+
+    if entity == coverage_gaps.low_measles_vaccine_coverage_first_dose:
+        paf_data['risk_id'] = entity.gbd_id
+        paf_data['coverage_gap_id'] = entity.gbd_id
+
     return paf_data
 
 
@@ -597,8 +609,7 @@ def _get_population_attributable_fraction(entities, location_ids):
             paf.extend([_compute_paf_for_special_cases(cause, entity, location_ids) for cause
                         in affected_causes if affected_causes])
             paf.extend([_compute_paf_for_special_cases(risk_factor, entity, location_ids) for risk_factor
-                        in affected_risk_factors if
-                        affected_risk_factors])
+                        in affected_risk_factors if affected_risk_factors])
         measure_data = pd.concat(paf)
 
     else:
@@ -650,7 +661,7 @@ def _get_exposure(entities, location_ids):
         SPECIAL = [coverage_gaps.low_measles_vaccine_coverage_first_dose]
         special_cases = set(SPECIAL).intersection(set(entities))
         if special_cases:
-            measure_data = gbd.get_exposures(risk_ids=special_cases, location_ids=location_ids)
+            measure_data = gbd.get_exposures(risk_ids=[s.gbd_id for s in special_cases], location_ids=location_ids)
             measure_data = _handle_special_coverage_gap_data(special_cases, measure_data, 0)
             measure_data = handle_exposure_from_gbd(measure_data)
             del measure_data['modelable_entity_id']
@@ -662,7 +673,7 @@ def _get_exposure(entities, location_ids):
             for entity in entities:
                 measure_data = gbd.get_auxiliary_data('exposure', 'coverage_gap', entity.name)
                 measure_data = measure_data[measure_data.location_id.isin(location_ids)]
-                measure_data['coverage_gap_id'] = np.NaN
+                measure_data['coverage_gap'] = entity.name
                 del measure_data['measure']
                 df.append(measure_data)
         exposure_data = pd.concat(df)
@@ -697,8 +708,8 @@ def _handle_special_coverage_gap_data(entities, measure_data, fill_value):
 
         coverage_gap_data = measure_data['coverage_gap_id'] == coverage_gap.gbd_id
         correct_age_groups = measure_data['age_group_id'].isin(good_age_groups)
-        coverage_gap_data['rei_id'] = np.NaN
-        coverage_gap_data['coverage_gap'] = coverage_gap.name
+        measure_data['rei_id'] = np.NaN
+        measure_data['coverage_gap'] = coverage_gap.name
         draw_cols = [f'draw_{i}' for i in range(1000)]
         measure_data.loc[coverage_gap_data & ~correct_age_groups, draw_cols] = fill_value
 
@@ -834,7 +845,7 @@ def get_covariate_estimates(covariates, locations):
 
 
 def get_location_ids_by_name():
-    return  {r.location_name: r.location_id for _, r in gbd.get_location_ids().iterrows()}
+    return {r.location_name: r.location_id for _, r in gbd.get_location_ids().iterrows()}
 
 
 def get_location_names_by_id():