+++ CHANGELOG.rst
@@ -1,3 +1,13 @@
+**1.0.4 - 09/28/18**
+
+ - Added measles covariates
+
+**1.0.3 - 09/25/18**
+
+ - Update for coverage gap exposure/rr/paf
+ - Add support for building artifacts for locations with spaces.
+ - Bugfixes around measles
+
 **1.0.2 - 09/12/18**
 
  - Add logging to artifact builder.